+++ spec/rails_helper.rb
@@ -27,6 +27,7 @@ Capybara.javascript_driver = :selenium_chrome_headless
 # require only the support files necessary.
 #
 # Dir[Rails.root.join('spec', 'support', '**', '*.rb')].each { |f| require f }
+Dir[Rails.root.join('spec/**/*_spec.rb')].each { |f| require f } # only need the nested spec files
 
 # Checks for pending migrations and applies them before tests are run.
 # If you are not using ActiveRecord, you can remove these lines.
